+++ src/main/java/edu/ie3/datamodel/utils/validation/UniquenessValidationUtils.java
@@ -11,7 +11,8 @@ import edu.ie3.datamodel.models.Entity;
 import edu.ie3.datamodel.models.UniqueEntity;
 import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.IdCoordinateInput;
-import edu.ie3.datamodel.models.result.ResultEntity;
+import edu.ie3.datamodel.models.result.CongestionResult;
+import edu.ie3.datamodel.models.result.ModelResultEntity;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
 import edu.ie3.datamodel.models.value.WeatherValue;
 import edu.ie3.datamodel.utils.Try;
@@ -27,8 +28,10 @@ public class UniquenessValidationUtils extends ValidationUtils {
   protected static final FieldSetSupplier<UniqueEntity> uuidFieldSupplier =
       entity -> Set.of(entity.getUuid());
   protected static final FieldSetSupplier<AssetInput> idFieldSupplier = e -> Set.of(e.getId());
-  protected static final FieldSetSupplier<ResultEntity> resultFieldSupplier =
+  protected static final FieldSetSupplier<ModelResultEntity> modelResultFieldSupplier =
       entity -> Set.of(entity.getTime(), entity.getInputModel());
+  protected static final FieldSetSupplier<CongestionResult> congestionResultFieldSupplier =
+      entity -> Set.of(entity.getTime(), entity.getSubgrid());
   protected static final FieldSetSupplier<MappingEntry> mappingFieldSupplier =
       entity -> Set.of(entity.participant());
   protected static final FieldSetSupplier<IdCoordinateInput> idCoordinateSupplier =
@@ -69,14 +72,25 @@ public class UniquenessValidationUtils extends ValidationUtils {
   }
 
   /**
-   * Checks the uniqueness of a collection of {@link ResultEntity}.
+   * Checks the uniqueness of a collection of {@link CongestionResult}.
    *
    * @param entities to be checked
    * @throws DuplicateEntitiesException if uniqueness is violated
    */
-  public static void checkResultUniqueness(Collection<? extends ResultEntity> entities)
+  public static void checkCongestionResultUniqueness(
+      Collection<? extends CongestionResult> entities) throws DuplicateEntitiesException {
+    checkUniqueness(entities, congestionResultFieldSupplier).getOrThrow();
+  }
+
+  /**
+   * Checks the uniqueness of a collection of {@link ModelResultEntity}.
+   *
+   * @param entities to be checked
+   * @throws DuplicateEntitiesException if uniqueness is violated
+   */
+  public static void checkModelResultUniqueness(Collection<? extends ModelResultEntity> entities)
       throws DuplicateEntitiesException {
-    checkUniqueness(entities, resultFieldSupplier).getOrThrow();
+    checkUniqueness(entities, modelResultFieldSupplier).getOrThrow();
   }
 
   /**